+++ src/app/models/review.py
@@ -11,8 +11,8 @@ class ReviewStatus(str, enum.Enum):
 
 class Review(Base):
     review_id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(BigInteger, ForeignKey("user.user_id"), nullable=False)
-    book_id = Column(BigInteger, ForeignKey("book.book_id"), nullable=False)
+    user_id = Column(Integer, ForeignKey("user.user_id"), nullable=False)
+    book_id = Column(Integer, ForeignKey("book.book_id"), nullable=False)
     
     rating = Column(Integer, nullable=False)
     content = Column(Text, nullable=False)
